[Refactor] #4622 baseitem-definition.cpp/h からbaseitem-key.cpp/h を分離した
