+++ Bakabakaband/Bakabakaband/Bakabakaband.vcxproj.filters
@@ -2523,6 +2523,9 @@
     <ClCompile Include="..\..\src\system\baseitem\baseitem-definition.cpp">
       <Filter>system\baseitem</Filter>
     </ClCompile>
+    <ClCompile Include="..\..\src\system\baseitem\baseitem-key.cpp">
+      <Filter>system\baseitem</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\..\src\combat\shoot.h">
@@ -5459,6 +5462,9 @@
       <Filter>system\enums</Filter>
     </ClInclude>
     <ClInclude Include="..\..\src\system\baseitem\baseitem-definition.h">
+      <Filter>system\baseitem</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\src\system\baseitem\baseitem-key.h">
       <Filter>system\baseitem</Filter>
     </ClInclude>
   </ItemGroup>
