+++ src/Makefile.am
@@ -943,6 +943,7 @@ hengband_SOURCES = \
 	system/terrain-type-definition.cpp system/terrain-type-definition.h \
 	system/gamevalue.h \
 	\
+	system/baseitem/baseitem-definition.cpp system/baseitem/baseitem-definition.h \
 	system/baseitem/baseitem-key.cpp system/baseitem/baseitem-key.h \
 	\
 	system/enums/game-option-page.h \
