+++ src/action/activation-execution.cpp
@@ -43,6 +43,7 @@
 #include "sv-definition/sv-ring-types.h"
 #include "system/artifact-type-definition.h"
 #include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-key.h"
 #include "system/floor-type-definition.h"
 #include "system/item-entity.h"
 #include "system/monster-entity.h"
