+++ src/action/open-close-execution.cpp
@@ -24,6 +24,7 @@
 #include "specific-object/chest.h"
 #include "status/bad-status-setter.h"
 #include "status/experience.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/dungeon-info.h"
 #include "system/floor-type-definition.h"
 #include "system/grid-type-definition.h"
