+++ src/artifact/artifact-info.cpp
@@ -9,6 +9,7 @@
 #include "object-enchant/tr-types.h"
 #include "smith/object-smith.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/item-entity.h"
 #include "util/bit-flags-calculator.h"
 #include "util/enum-converter.h"
