+++ src/flavor/flavor-util.cpp
@@ -5,6 +5,7 @@
 #include "object/tval-types.h"
 #include "sv-definition/sv-food-types.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/item-entity.h"
 #include "util/string-processor.h"
 #include <sstream>
