+++ src/flavor/named-item-describer.cpp
@@ -11,6 +11,7 @@
 #include "object-enchant/tr-types.h"
 #include "perception/object-perception.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "util/bit-flags-calculator.h"
