+++ src/market/arena-entry.cpp
@@ -8,7 +8,7 @@
 #include "sv-definition/sv-scroll-types.h"
 #include "sv-definition/sv-staff-types.h"
 #include "sv-definition/sv-wand-types.h"
-#include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-key.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/monster-race-info.h"
 #ifndef JP
