+++ src/market/bounty-prize-table.cpp
@@ -2,7 +2,7 @@
 #include "object/tval-types.h"
 #include "sv-definition/sv-potion-types.h"
 #include "sv-definition/sv-scroll-types.h"
-#include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-key.h"
 
 const std::vector<BaseitemKey> prize_list = {
     { ItemKindType::POTION, SV_POTION_CURING },
