+++ src/mind/mind-hobbit.cpp
@@ -1,6 +1,7 @@
 #include "mind/mind-hobbit.h"
 #include "floor/floor-object.h"
 #include "sv-definition/sv-food-types.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/item-entity.h"
 #include "system/player-type-definition.h"
 #include "view/display-messages.h"
