+++ src/monster-floor/monster-death.cpp
@@ -29,6 +29,7 @@
 #include "sv-definition/sv-scroll-types.h"
 #include "system/angband-system.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/building-type-definition.h"
 #include "system/dungeon-info.h"
 #include "system/enums/monrace/monrace-id.h"
