+++ src/monster/monster-processor.cpp
@@ -62,6 +62,7 @@
 #include "spell/summon-types.h"
 #include "sv-definition/sv-junk-types.h"
 #include "system/angband-system.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/enums/monrace/monrace-id.h"
 #include "system/floor-type-definition.h"
 #include "system/grid-type-definition.h"
