+++ src/player/player-damage.cpp
@@ -54,6 +54,7 @@
 #include "status/base-status.h"
 #include "status/element-resistance.h"
 #include "sv-definition/sv-junk-types.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/building-type-definition.h"
 #include "system/dungeon-info.h"
 #include "system/floor-type-definition.h"
