+++ src/store/articles-on-sale.cpp
@@ -16,7 +16,7 @@
 #include "sv-definition/sv-staff-types.h"
 #include "sv-definition/sv-wand-types.h"
 #include "sv-definition/sv-weapon-types.h"
-#include "system/baseitem/baseitem-definition.h"
+#include "system/baseitem/baseitem-key.h"
 
 /*!
  * @brief 店舗で常時販売するオブジェクトを定義する
