+++ src/store/rumor.cpp
@@ -57,7 +57,7 @@ static std::string bind_rumor_name(std::string_view base, std::string_view item_
  * @return トークン読み込み成否 とトークン群の配列
  * @todo tmp_tokensを使わず単なるsplitにすればもっと簡略化できそう
  */
-std::pair<bool, std::vector<std::string>> get_rumor_tokens(std::string rumor)
+static std::pair<bool, std::vector<std::string>> get_rumor_tokens(std::string rumor)
 {
     constexpr auto num_tokens = 3;
     char *tmp_tokens[num_tokens];
