+++ src/system/floor-type-definition.cpp
@@ -5,6 +5,7 @@
 #include "monster/monster-timed-effects.h"
 #include "system/angband-system.h"
 #include "system/artifact-type-definition.h"
+#include "system/baseitem/baseitem-definition.h"
 #include "system/dungeon-info.h"
 #include "system/enums/grid-count-kind.h"
 #include "system/gamevalue.h"
